Major rewrite of slha-derive
Instead of using fragile special cases in the derive code to detect
vectors and options, use the new trait 'WrappedBlock' to handle these
cases.
The new code is supposed to handle attributes and already allows to
rename blocks.
All the internal code that is necessarily public to allow for
slha-derive to work was moved in the internal directory.