Big rewrite of the Slha struct and API changes
The checks for duplicate blocks have been moved from the first parsing
stage in `Slha::parse` to the `get_block` functions.
A `get_blocks_unchecked` method was added to obtain all blocks (with a
given name) without any checks for duplicates.

Internally, the `BlockScale` enum was replaced by a new public type
`RawBlock` that contains an optional scale and the `Line`s belonging to
a block. Later on functions will be added to obtain these unparsed blocks
from an `Slha` object.