Introduce generic block headers
The old Block, BlockStr and BlockSingle have been expressed using the
generic BlockExtra and BlockSingleExtra that can contain any type of
additional information in the header as long as its rust type implements
Parseable.
This is a preparation for also allowing blocks with pdg ids in the
header, i.e. the QNUMBERS block.